ensure java enum used for enum parameters #273
